Migrate CondenseText from textConstants to textEffect. Update condenseText helper with a condensing threshold and input length assertion.
